% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ate.R
\name{get_ate}
\alias{get_ate}
\title{get_ate}
\usage{
get_ate(dag_model, ...)
}
\arguments{
\item{dag_model}{An object of either class "paramteric_dag_model" or
"non_paramteric_dag_model".}

\item{...}{Additional arguments to dispatched method.}
}
\value{
A data.frame with 3 columns:
\enumerate{
\item From: The baseline treatment value.
\item To: The value a treatment was changed to.
\item ATE: The average treatment effect of changing the treatment
value from "From" to "To.
}
}
\description{
\code{get_ate} is a generic function used to
get average treatment effects (\eqn{ATE}) in a DAG model.
}
\seealso{
\code{\link{get_ate.non_parametric_dag_model}}, \code{\link{get_ate.parametric_dag_model}}
}
