% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_dag_model.R
\name{non_parametric_dag_model}
\alias{non_parametric_dag_model}
\title{Specify a non parametric DAG model}
\usage{
non_parametric_dag_model(dag, data)
}
\arguments{
\item{dag}{An object of class "dagitty" representing the DAG.}

\item{data}{A data.frame to be used for fitting the GAMs (no missing values allowed).}
}
\value{
An object of class "non_parametric_dag_model" which is essentially a list containing the following
elements:
\enumerate{
\item dag: The model DAG
\item gam_fits: All smoothing functions and other information required for simulating
new datasets from the specified dag model
}
}
\description{
\code{non_parametric_dag_model} is used to specify non paramteric
DAG models. The non parametric node functions are obtained by fitting generalized
additive models to a supplied (usually "real life") dataset.
The resulting obejct of class "non_parametric_dag_model" has methods to simulate
observations from the model specified and get \eqn{ATE}s.
}
\details{
Every node \eqn{j} in a non parametric DAG model is simulated from a
non paramteric function of the node parents \code{PA_{j}}:

\deqn{f_j(PA_j) = g(\sum_{i \in PA_j}s_{ij}(x_i) + \epsilon_j)}

The smooth functions \eqn{s_{ij}(.)} are obtained by fitting generalized
additive model to the node (as a reponse) and it's parents (as predictors).
}
\examples{
require(carData); require(dagitty)
data("GSSvocab")
GSSvocab <- GSSvocab[complete.cases(GSSvocab), ]
g <- dagitty("dag {
             ageGroup [pos=\\"0,0\\"]
             vocab [pos=\\"1,-1\\"]
             nativeBorn [pos=\\"2,-2\\"]
             educ [pos=\\"3,-1\\"]
             gender [pos=\\"4,0\\"]
             nativeBorn -> educ
             nativeBorn -> vocab
             educ -> vocab
             gender -> educ
             ageGroup -> vocab
             }")
plot(g)
non_param_dag_model <- non_parametric_dag_model(dag = g, data = GSSvocab)
sim_data <- sim_mixed_dag(non_param_dag_model, N = 30000)
boxplot(vocab ~ nativeBorn, data = sim_data)
# verify distribtion in simulated dataset looks similar
boxplot(vocab ~ nativeBorn, data = GSSvocab)
}
\seealso{
\code{\link{parametric_dag_model}} for parametric DAG model specification. Methods for
the non_parametric_dag_model class include \code{\link{sim_mixed_dag.non_parametric_dag_model}} for
simulating datasets and \code{\link{get_ate.non_parametric_dag_model}} for getting \eqn{ATE}s.
}
