% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mixed_dag.non_parametric_dag_model.R
\name{sim_mixed_dag.non_parametric_dag_model}
\alias{sim_mixed_dag.non_parametric_dag_model}
\title{Simulate a dataset from a non parametric DAG model}
\usage{
\method{sim_mixed_dag}{non_parametric_dag_model}(dag_model, N = 1000,
  ...)
}
\arguments{
\item{dag_model}{An object of class "non_parametric_dag_model".}

\item{N}{Number of observations to simulate.}

\item{...}{Additional arguments}
}
\value{
A dataset simulated from the input \code{dag_model}.
}
\description{
Simulate a dataset from a non parametric DAG model.
}
\examples{
require(carData); require(dagitty)
data("GSSvocab")
GSSvocab <- GSSvocab[complete.cases(GSSvocab), ]
g <- dagitty("dag {
             ageGroup [pos=\\"0,0\\"]
             vocab [pos=\\"1,-1\\"]
             nativeBorn [pos=\\"2,-2\\"]
             educ [pos=\\"3,-1\\"]
             gender [pos=\\"4,0\\"]
             nativeBorn -> educ
             nativeBorn -> vocab
             educ -> vocab
             gender -> educ
             ageGroup -> vocab
             }")
plot(g)
non_param_dag_model <- non_parametric_dag_model(dag = g, data = GSSvocab)
sim_data <- sim_mixed_dag(dag_model = non_param_dag_model, N = 30000)
boxplot(vocab ~ nativeBorn, data = sim_data)
# verify distribtion in simulated dataset looks similar
boxplot(vocab ~ nativeBorn, data = GSSvocab)
}
\seealso{
\code{\link{sim_mixed_dag.parametric_dag_model}} for parametric dag simulation.
}
