% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mixed_dag.parametric_dag_model.R
\name{sim_mixed_dag.parametric_dag_model}
\alias{sim_mixed_dag.parametric_dag_model}
\title{Simulate a dataset from a parametric DAG model}
\usage{
\method{sim_mixed_dag}{parametric_dag_model}(dag_model, N = 1000, ...)
}
\arguments{
\item{dag_model}{An object of class "parametric_dag_model".}

\item{N}{Number of observations to simulate.}

\item{...}{Additional arguments.}
}
\value{
A dataset simulated from the input \code{dag_model}.
}
\description{
Simulate a dataset from a parametric DAG model.
}
\examples{
require(dagitty)
g <- dagitty("dag {
sales [pos=\\"0,0\\"]
             mkt [pos=\\"2,0\\"]
             comp [pos=\\"1,1\\"]
             visits [pos=\\"1,0\\"]
             visits -> sales
             mkt -> visits
             comp -> mkt
             comp -> sales
             }")

plot(g)

param_dag_model <- parametric_dag_model(
  dag = g,
  f.args = list(
    sales = list(betas = list(visits = 0.3, comp = -0.9)),
    visits = list(betas = list(mkt = 0.5)),
    mkt = list(betas = list(comp = 0.6))
  )
)

sim_data <- sim_mixed_dag(dag_model = param_dag_model)

plot(sim_data$mkt, sim_data$sales) # confounded relation
}
\seealso{
\code{\link{sim_mixed_dag.non_parametric_dag_model}} for non parametric dag simulation.
}
