\name{simMixedDAG-package}
\alias{simMixedDAG-package}
\alias{simMixedDAG}
\docType{package}
\title{
\packageTitle{simMixedDAG}
}
\description{
\packageDescription{simMixedDAG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simMixedDAG}
\packageIndices{simMixedDAG}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simMixedDAG}

Maintainer: \packageMaintainer{simMixedDAG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
