% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_dag_model.R
\name{non_parametric_dag_model}
\alias{non_parametric_dag_model}
\title{Specify a non parametric DAG model}
\usage{
non_parametric_dag_model(dag, data)
}
\arguments{
\item{dag}{An object of class "dagitty" representing the DAG.}

\item{data}{A data.frame to be used for fitting the GAMs (no missing values allowed).}
}
\value{
An object of class "non_parametric_dag_model" which is essentially a list containing the following
elements:
\enumerate{
\item dag: The model DAG
\item gam_fits: All smoothing functions and other information required for simulating
new datasets from the specified dag model
}
}
\description{
\code{non_parametric_dag_model} is used to specify non paramteric
DAG models. The non parametric node functions are obtained by fitting generalized
additive models to a supplied (usually "real life") dataset.
The resulting obejct of class "non_parametric_dag_model" has methods to simulate
observations from the model specified and get \eqn{ATE}s.
}
\details{
Every node \eqn{j} in a non parametric DAG model is simulated from a
non paramteric function of the node parents \code{PA_{j}}:

\deqn{f_j(PA_j) = g(\sum_{i \in PA_j}s_{ij}(x_i) + \epsilon_j)}

The smooth functions \eqn{s_{ij}(.)} are obtained by fitting generalized
additive model to the node (as a reponse) and it's parents (as predictors).
}
\examples{
require(dagitty)
g <- dagitty("dag {
sales [pos=\\"0,0\\"]
             mkt [pos=\\"2,0\\"]
             comp [pos=\\"1,1\\"]
             visits [pos=\\"1,0\\"]
             visits -> sales
             mkt -> visits
             comp -> mkt
             comp -> sales
             }")

plot(g)

param_dag_model <- parametric_dag_model(
  dag = g,
  f.args = list(
    sales = list(betas = list(visits = 0.3, comp = -0.9)),
    visits = list(betas = list(mkt = 0.5)),
    mkt = list(betas = list(comp = 0.6))
  )
)

sim_data <- sim_mixed_dag(param_dag_model)

plot(sim_data$mkt, sim_data$sales) # confounded relation
}
\seealso{
\code{\link{parametric_dag_model}} for parametric DAG model specification. Methods for
the non_parametric_dag_model class include \code{\link{sim_mixed_dag.non_parametric_dag_model}} for
simulating datasets and \code{\link{get_ate.non_parametric_dag_model}} for getting \eqn{ATE}s.
}
